feat/refactor-components: nest match-tiles folder under dashboard folder
